+++ Tests/SPMBuildCoreTests/PluginInvocationTests.swift
@@ -194,6 +194,7 @@ class PluginInvocationTests: XCTestCase {
             builtToolsDir: builtToolsDir,
             buildEnvironment: BuildEnvironment(platform: .macOS, configuration: .debug),
             toolSearchDirectories: [UserToolchain.default.swiftCompilerPath.parentDirectory],
+            pkgConfigDirectories: [],
             pluginScriptRunner: pluginRunner,
             observabilityScope: observability.topScope,
             fileSystem: fileSystem
@@ -886,6 +887,7 @@ class PluginInvocationTests: XCTestCase {
                     builtToolsDir: builtToolsDir,
                     buildEnvironment: BuildEnvironment(platform: .macOS, configuration: .debug),
                     toolSearchDirectories: [UserToolchain.default.swiftCompilerPath.parentDirectory],
+                    pkgConfigDirectories: [],
                     pluginScriptRunner: pluginScriptRunner,
                     observabilityScope: observability.topScope,
                     fileSystem: localFileSystem
@@ -1188,6 +1190,7 @@ class PluginInvocationTests: XCTestCase {
                 builtToolsDir: builtToolsDir,
                 buildEnvironment: BuildEnvironment(platform: .macOS, configuration: .debug),
                 toolSearchDirectories: [UserToolchain.default.swiftCompilerPath.parentDirectory],
+                pkgConfigDirectories: [],
                 pluginScriptRunner: pluginScriptRunner,
                 observabilityScope: observability.topScope,
                 fileSystem: localFileSystem